Separate the classes using 3 files
