+++ person.rb
@@ -1,0 +1,23 @@
+class Person
+    attr_accessor :name, :age
+    attr_reader :id
+  
+    def initialize(id, age, name = 'Unknown', parent_permission: true)
+      @id = id
+      @name = name
+      @age = age
+      @parent_permission = parent_permission
+    end
+  
+    # Name and age setters
+  
+    def can_use_services?
+      of_age? || @parent_permission
+    end
+  
+    private
+  
+    def of_age?
+      @age >= 18
+    end
+  end
